flow: complete code (check approval cond)

+++ backend/middlewares/flow/src/dto/flow_state_dto.rs
@@ -171,7 +171,7 @@ pub struct FlowStateCountersignConf {
     /// 类型
     pub kind: FlowStateCountersignKind,
     /// 多数人通过比例
-    pub most_percent: Option<i8>,
+    pub most_percent: Option<usize>,
     /// 审批人权限配置
     pub guard_custom_conf: Option<FlowGuardConf>,
     /// 指定人通过即通过
